+++ src/main/scala/edu/ie3/simona/service/primary/PrimaryServiceProxy.scala
@@ -420,48 +420,29 @@ object PrimaryServiceProxy {
   ): Try[ActorRef[Message]] = {
     val valueClass = metaInformation.getColumnScheme.getValueClass
 
-    val workerRef = classToWorkerRef(metaInformation.getUuid.toString)
     toInitData(
       metaInformation,
       simulationStart,
       primaryConfig,
       valueClass,
-    ) match {
-      case Success(initData) =>
-        workerRef ! Create(
+    ).recoverWith { cause =>
+      Failure(
+        new InitializationException(
+          "Unable to build init data for worker. Kill the uninitialized worker. Goodbye my friend!",
+          cause,
+        )
+      )
+    }.map { initData =>
+      ctx.spawn(
+        PrimaryServiceWorker(
+          scheduler,
           initData,
           ScheduleLock.singleKey(ctx, scheduler, INIT_SIM_TICK),
-        )
-        Success(workerRef)
-      case Failure(cause) =>
-        ctx.stop(workerRef)
-        Failure(
-          new InitializationException(
-            "Unable to build init data for worker. Kill the uninitialized worker. Goodbye my friend!",
-            cause,
-          )
-        )
+        ),
+        metaInformation.getUuid.toString,
+      )
     }
   }
-
-  /** Build a primary source worker and type it to the foreseen value class to
-    * come
-    *
-    * @param timeSeriesUuid
-    *   Uuid of the time series the actor processes
-    * @return
-    *   The [[ActorRef]] to the spun off actor
-    */
-  private[service] def classToWorkerRef(
-      timeSeriesUuid: String
-  )(using
-      scheduler: ActorRef[SchedulerMessage],
-      ctx: ActorContext[Message],
-  ): ActorRef[Message] =
-    ctx.spawn(
-      PrimaryServiceWorker(scheduler),
-      timeSeriesUuid,
-    )
 
   /** Building proper init data for the worker
     *